Work in progress on #1 This commit was completed in collaboration between Alan and Lawrence.
Fixed the idea of the structs for the decoded memory, now
have a consistant underlying setup on which to base our
output.

We've made use of some of the C preprocessor directives
in order to DRY up the files along with enhancing the idea
of our our hacked together polymorphism/inheritance.

With any luck, should get something running in about and hour,
from which point we can catch up with testing edge cases!